% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib.R
\name{contrib}
\alias{contrib}
\title{Contributions aux mesures de risque VaR et TVaR}
\usage{
contrib(u, fi)
}
\arguments{
\item{u}{Niveau de confiance.}

\item{fi}{Matrice où chaque colonne est la fonction de masse
de probabilité de \eqn{M_i}.}
}
\value{
Une liste contenant :
\describe{
\item{V.a. S}{VaR et TVaR de \eqn{S}.}
\item{CVaR}{Vecteur des contributions à la VaR.}
\item{CTVaR}{Vecteur des contributions à la TVaR.}
}
}
\description{
Calcule la contribution de la v.a. \eqn{M_i} à la VaR et TVaR
de \eqn{S = M_1 + \cdots + M_d}.
}
\examples{
k <- 0:1000
n <- c(200, 300)
q <- c(0.15, 1/15)
u <- 0.99

fm1 <- dbinom(k, n[1], q[1])
fm2 <- dbinom(k, n[2], q[2])
fmi <- cbind(fm1, fm2)
contrib(u = 0.95, fmi)

}
