% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloi_lim_fran.R
\name{dloi_lim_fran}
\alias{dloi_lim_fran}
\title{Fonction de masse de probabilité avec franchise et limite}
\usage{
dloi_lim_fran(d, u, k, f)
}
\arguments{
\item{d}{Un scalaire numérique représentant la limite maximale de couverture.}

\item{u}{Un scalaire numérique représentant le déductible.}

\item{k}{Un vecteur numérique représentant le domaine de la variable aléatoire.}

\item{f}{Fonction de masse de probabilité. Doit sommer à 1.}
}
\value{
Un vecteur numérique représentant la fonction de masse de probabilité
de la variable aléatoire transformée \code{min(max(k - fran, 0), lim)}.
}
\description{
Calcule la fonction de masse de probabilité d'une v.a. discrète
après application d'une franchise et d'une limite.
}
\examples{
k <- 0:1000
r <- 2.5
q <- 0.7
d <- 2
u <- 8
fx <- dnbinom(k, r, q)
dloi_lim_fran(d, u, k, fx)
}
