% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvar.R
\name{TVaR}
\alias{TVaR}
\title{Tail Value at Risk (TVaR)}
\usage{
TVaR(u, k, f)
}
\arguments{
\item{u}{Un scalaire numérique représentant le niveau de confiance (par
exemple, 0.95 pour 95\%).}

\item{k}{Un vecteur numérique représentant le domaine variable aléatoire.}

\item{f}{Fonction de masse de probabilité. Doit sommer à 1.}
}
\value{
Un scalaire numérique représentant la TVaR.
}
\description{
Calcule la TVaR à un niveau de confiance donné à partir
de la fonction de masse de probabilité pour une v.a. discrète.
}
\examples{
f <- c(0.1, 0.2, 0.4, 0.2, 0.1)
k <- 0:4
TVaR(u = 0.95, k, f)

}
